pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medvqa-probe"
version = "0.1.0"
description = "Internal-state hallucination classifier for MedVQA models"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.36",
    "datasets>=2.14",
    "Pillow>=10.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-cov"]

[tool.setuptools.packages.find]
include = ["medvqa_probe*"]
